Move logic to the existing router
