+++ tools/webpack/interactivity.js
@@ -19,7 +19,6 @@ module.exports = {
 	entry: {
 		index: './packages/interactivity',
 		router: './packages/interactivity-router',
-		'full-page-router': './packages/interactivity-router/src/full-page.js',
 		navigation: './packages/block-library/src/navigation/view.js',
 		query: './packages/block-library/src/query/view.js',
 		image: './packages/block-library/src/image/view.js',
